Add option to import centers from ROI manager
If the ROIs are polygons, centroids are assumed as the centers.
In case of point ROI, they are directly imported. The idea is to
be able to use this plugin with other segmentation methods. (ie.
I use "analyze particles" to find spherical small objects and
use the balloons to encapsulate them.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>