POM: fix copyright owners
Well, not exactly "fixed" since "Fiji developers" is tautological.
But trying to build an accurate list of copyright owners for Fiji
is a substantial endeavor. Git already knows who contributed, and
documenting it precisely in the license headers brings us no
immediate additional technical benefit, so for now we punt.